baselines/Baran/Error-Correction-at-Scale/benchmarks/Baran_Experiments/raha/requirements.txt
numpy==1.22.4
pandas==2.1.0
scipy==1.10.0
scikit-learn==1.2.2
matplotlib==3.7.1
prettytable==0.7.2
mwparserfromhell==0.5.4
beautifulsoup4==4.6.0
py7zr==0.20.2
-e .
hydra-core==1.3
requests==2.31.0
flask==2.2.2
python-dotenv==1.0.0